server/requirements.txt
Flask==1.1.2
Flask_Cors==3.0.8
pdf2image==1.16.0
PyPDF2==1.26.0
python-dotenv==0.19.2
Pillow~=8.3.2